[BE] - FIX : DonationList 수정

+++ birdmeal/src/main/java/com/backend/birdmeal/dto/ResponseDonationDto.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ResponseDonationDto {
-    private int donationSeq;
+    private long donationSeq;
     private String userNickname;
     private int donationPrice;
     private String donationDate;
